[todolist > work-in-progress note taking] Added functionality: A "workitem" snippet can be triggered inside a markdown file.
I've always struggled with keeping track of the link to my current
task, as well as logging the rest requests/answers I've tried in the
past so as to repeat them. Additionally, I've always wanted to write
commit messages that were more than a succint description of the added
feature or nature of the code modification.

Using a markdown file will help me keep track of all the different
resources I'm using, as well as providing the ability to re-play past
queries that worked in a past context.